requirements.txt
kubernetes==29.0.0
boto3==1.34.11
pretty-html-table==0.1.1
requests==2.31.0
pyYAML==6.0.1